// modules/redux/user/user.enum.ts
export enum UserActionTypes {
  EMAIL_SIGN_IN_START = 'EMAIL_SIGN_IN_START',
  SIGN_IN_SUCCESS = 'SIGN_IN_SUCCESS',
  SIGN_UP_START = 'SIGN_UP_START',
  SIGN_UP_SUCCESS = 'SIGN_UP_SUCCESS',
  SIGN_OUT_START = 'SIGN_OUT_START',
  SIGN_OUT_SUCCESS = 'SIGN_OUT_SUCCESS',
  GET_RESET_PASSWORD_EMAIL_START = 'GET_RESET_PASSWORD_EMAIL_START',
  GET_RESET_PASSWORD_EMAIL_SUCCESS = 'GET_RESET_PASSWORD_EMAIL_SUCCESS',
  RESET_PASSWORD_START = 'RESET_PASSWORD_START',
  RESET_PASSWORD_SUCCESS = 'RESET_PASSWORD_SUCCESS',
  ONBOARD_USER_START = 'ONBOARD_USER_START',
  UPDATE_USER_START = 'UPDATE_USER_START',
  UPDATE_USER_SUCCESS = 'UPDATE_USER_SUCCESS',
  LOAD_INITIAL_USER_DATA = 'LOAD_INITIAL_USER_DATA',
  GOOGLE_SIGN_IN_START = 'GOOGLE_SIGN_IN_START',
  GOOGLE_SIGN_IN_SUCCESS = 'GOOGLE_SIGN_IN_SUCCESS',
  SET_USER_OPERATION_ERROR = 'SET_USER_OPERATION_ERROR',
  SET_HAS_LOCATION = 'SET_HAS_LOCATION',
  SET_LOCATION_COOKIE_START = 'SET_LOCATION_COOKIE_START',
  SET_AUTH_MODAL = 'SET_AUTH_MODAL',
  SET_ACCESS_TOKEN = 'SET_ACCESS_TOKEN',
}

export enum UserEnum {
  User = 'user',
  Data = 'data',
  Error = 'error',
  Loading = 'loading',
  HasLocation = 'hasLocation',
  AuthModal = 'authModal',
  AccessToken = 'accessToken',
}

export enum AuthModalEnum {
  Options = 'options',
  SignIn = 'signIn',
  SignUp = 'signUp',
  ResetPassword = 'resetPassword',
  Onboarding = 'onboarding',
}